찬구 모달 코드 개선
