+++ client/src/components/Layout/HeaderFriendModal/FriendModal.style.tsx
