+++ client/src/components/Post/Post.style.tsx
@@ -89,6 +89,10 @@ export const VisibilityIcon = styled(Icon)({
 });
 
 export const Image = styled('img')({
+  position: 'absolute',
+  backgroundRepeat: 'no-repeat',
+  backgroundSize: 'cover',
+  backgroundPosition: 'center',
   width: '100%',
   height: '100%',
 });
